+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Instance explorer is refreshed every 15 seconds.
 - Refresh intervall can be configured via the `lxd.refreshInterval` property.
 - LXD images explorer.
+- LXD networks explorer.
 
 ### Fixed